Add files via upload
1.1.	Introduction
            When the novel coronavirus (Covid-19) pandemic emerges, the spread of the virus has left public keep anxiety if they do not have any effective cure. The World Health Organization (WHO) has declared Covid-19 as a pandemic due to the increase in the number of cases reported around the world. To contain the pandemic, many countries have implemented a lockdown where the government enforced that the citizens to stay at home during this critical period. The public health bodies such as the Centers for Disease Control and Prevention (CDC) had to make it clear that the most effective way to slow down the spread of Covid-19 is by avoiding close contact with other people. To flatten the curve on the Covid-19 pandemic, the citizens around the world are practicing physical distancing. To implement social distancing, group activities and
congregations such as travel, meetings, gatherings, workshops, praying had been banned during the quarantine period. The people are encouraged to use phone and email to manage and conduct events as much as possible to minimize the person-to-person contact. To further contain the spread of the virus, people are also informed to perform hygiene measures such as frequently washing hands, wearing mask and avoiding close contact with people who are ill. However, there is a difference between knowing what to do to reduce the transmission of the virus and putting them into practice. The world has not yet fully recover from this pandemic and the vaccine that can effectively treat Covid-19 is yet to be discovered. However, to reduce the impact of the pandemic on the country’s economy, several governments have allowed a limited number of economic activities to be  resumed once the number of new cases of Covid-10 has dropped below a certain level. As these countries cautiously restarting their economic activities, concerns have emerged regarding workplace safety in the new post-Covid-19 environment. To reduce the possibility of infection, it is advised that people should avoid any person-to-person contact such as shaking hands and they should maintain a distance of at least 1 meter from each other. In Malaysia, the Ministry of Health Malaysia (MOHM) has recommended several disease prevention measures for workplaces, individuals, and families at home, schools, childcare centres, and senior living facilities. These measures include implementing social distancing measures, increasing physical space between workers at the workplace, staggering work schedules, decreasing social contacts in the workplace, limiting large work-related gatherings, limiting non-essential work travel, performing regular health checks of staff and visitors entering buildings, reducing physical activities especially for organizations that have staff in the high-risk category, and conducting company events or activities online. Individuals, communities, businesses, and healthcare organizations are all part of a community with their responsibility to mitigate the spread of the Covid-19 disease. In reducing the impact of this coronavirus pandemic,
practicing social distancing and self-isolation have been deemed as the most effective ways to break the chain of infections after restarting the economic activities. In fact, it has been observed that there are many people who are ignoring public health measures, especially with respect to social distancing. It is understandable that given the people’s excitement to start working again, they sometimes tend to forget or neglect the implementation of social distancing. Hence, this work aims to facilitate the enforcement of social distancing by providing automated detection of social distance violation in workplaces and public areas using a deep learning model. In the area of machine learning and computer vision, there are different methods that can be used for object detection. These methods can also be applied to detect the social distance between people. The following points summarizes the main components of this approach:
•	Deep learning has gained more attention in object detection was used for human detection purposes.
•	Develop a social distancing detection tool that can detect the distance between people to keep safe.
•	Evaluation of the classification results by analyzing real-time video streams from the camera.
In the fight against the COVID-19 pandemic, social distancing, contactless meetings, wearing mask all are proven to be very effective measure to slow down the spread of the virus. Social distancing is the one of the best way to prevent from virus, people asked to keep distance of at least 6 feet from other to stop the spread of a virus.
Social distancing is a method used to control the spread of contagious diseases. It implies that people physically distance themselves from one another, reducing close contact, and thereby reducing the spread of a contagious disease (such as the COVID-19 Disease).
We build a social distance analyzer using computer vision, python, and object detection using YOLO algorithm. YOLO (You Only Look Once) real-time object detection algorithm, which is one of the most effective object detection algorithms that also encompasses many of the most innovative ideas coming out of the computer vision research community. We can use a OpenCV and YOLO to analyze or monitor whether the people are in distance or not.


 
1.2. Feature
•	Object detection using the YOLO model to detect only people in a video stream.
•	Computes the pairwise distances between all detected people.
•	Based on the computed distances, we determine whether social distancing rule is being violated or not.

1.3.	 OpenCV Library
OpenCV is the huge open-source library for the computer vision, machine learning, and image processing and now it plays a major role in real-time operation which is very important in today’s systems. By using it, one can process images and videos to identify objects, faces, or even handwriting of a human. When it integrated with various libraries, such as Numpy, python is capable of processing the OpenCV array structure for analysis. To identify image pattern and its various features we use vector space and perform mathematical operations on these features.